Python Course For Machine Learning and Data Science.
python is a popular language in the fields such as Data Science, Machine Learning. the reason behind it is that most people who are data scientists or Machine Learning professionals need not to be from a Computer Science background, most of the people are statistians, or Economists, etc. and as python is easy to understand, as well as to learn, python is their first preferance. and Hey, so as ours! Language such as R is also popular in this field and there has been a debate all the time about which one is better. but dont get into it, both languages are similarly powerfull to apply all kinds of ML and DL algorithms.
This Course is designed by Yash Balwir for AI Schools.